Working pipeline with new kwyk dataset, 2/51/107 classes, Segformer model, no augmentations
1. Add (temporary) flag to indicate whether to use new kwyk dataset or old datasets

2. Use brain_utils.mapping in metrics to convert from freesurfer label counts to nr_of_classes

3. Use brain_utils.mapping in dataset.py to convert labels from freesurfer labels to nr_of_classes

4. Modify get_data_loader() to return slice shape and feed into Segformer

+++ TissueLabeling/training/trainer.py
@@ -102,7 +102,7 @@ class Trainer:
             print(f"Process {self.fabric.global_rank}, batch {i}")
 
             self.optimizer.zero_grad()
-            probs = self.model(image)
+            probs = self.model(image.to(torch.float32))
             loss = self.loss_fn(mask.long(), probs)
             self.fabric.backward(loss)
             self.optimizer.step()
